Initial commit
This was my blog from 2002 - ...2006 I think.
I wrote this from scratch. It was the first significant thing I wrote for myself.
It doesn't work with the latest version of PHP - the functions mysql_* have been deprecated.
A initial review has been made to remove any files that look sensitive.